Added Files to Folders
Therefore I changed the pathways of nearly every <a> link in order for the pathways to work. I did this because I want to make it easier to see the files I am working on when I am in Atom.

Plus I worked on a Sample Form, but it's garbage. I have to look up the attributes of each element and sample it one by one. Right now... just look at it.