Minor Details
I added a line break between the About label and the text area, because the text area can be dragged and it forces the About to get ugly and go under it or move in all ways.

I also added a phone number placeholder for consistency. All it does now is show you a sample number.