Added a blog
I wanted to add a professional blog, because eventually my content *will* be out there online, with SEO and all. It won't hurt to have fun blogs, but for some reason I also feel compelled to add more "professional" blogs to my website.

It would help in that, I don't want to be a blogger but, I do want to see when certain additions to my website increase my website traffic. It would be nice to have content worth visiting, so that's the logic behind blog4