Improved a11y
Installed a screenreader as well as a couple extensions to understand readability.

I made the About -> About Me, because a JavaScript tool mentioned that About was not descriptive enough.

I also renamed my Portfolio links to at least say Tinker Project or Portfolio Sample 1 because Sample 1 was not clear for a screen reader.

Additionally, I added a section to my Portfolio table of contents. The Tinker Projects list item now leads to another table of contents. This page will include projects I work on, before they are ready for my final drafts.