Commented Out Sub Menus
I commented out portfolio and blog nested lists within each file.

The reason I did this was because there is no need to hover and see every blog (what if there's 100?!?!) or project from the navigation in the header.

For now, I need to figure out how to center a logo within a nav that is styled with grid and flex. But for now, rest.