Several Updates
Added Coke CSS to experiment with a separate style sheet.

Added the Logo-Right item (j9e image) to pages 4,5 and 6.

I changed the name of a link in portfolio to make it match.

I changed the name of Tinker 6 so that it better reflected what the project was.

Lastly, I experimented a bit with media queries, but my logo still looks bad since its image isn't proper format.

I made 9 divs and will experiment with grid the next time I have free time.