Added Coca-Cola Images
I fixed my Coca-Cola gap issue but what I am doing is a work-around, not the best solution. There is a margin on each item, same % in all four corners, making double the distance vertically between the 1st and 2nd and 2nd to last and last items.

I wonder if I should try recreating this layout with a vertical flex column? Or maybe just work on media queries for now, get back to the CSS Critters game, then rebuild it later.