Positioning FIles
I deleted several files in that I renamed them to other files.

I created a sticky nav bar but it isn't finished. I also plan on adding media queries to this so that I can make it as responsive as can be without there being a hamburger menu.

Will test on laptop later to make sure things pushed and pulled correctly.