Hamburger Media Queries
I used media queries to only make the sidebar navigation appear on mobile devices ranging between the Galaxy 5 and the iPad Pro.

I also expanded the navigation list to nearly 100% of the viewport width to see how that looked on mobile.

I also changed the toggle nav height so that it was even with the header, regardless of header height.