Travery's Hamburger Tutorial
I already doctored it a little bit but there's still a lot of studying to be done.

What I have to do is customize it to make it completely mine. That way I can integrate it with the other nav bar I've done, make it exactly mine and not just a mindless copy, then I can integrate it on the full website.