Footer, Blog, Spacing, Attribution
I standardized my footer by using flaticons but also changed the spacing on larger widths.

I deleted an <hr> and some superfluous spacing on the bottom of blog photos.

I deleted the attribution to phatsoup because when I went to verify it was required I did not find the icons used on that website, so the attribution was not said to be required.

I also duplicated the index file, as well as the mediaquery.css file where I originally made the icons.

I need to name files better.