Footer Updates
I made sure the heading was below the header on all device sizes.

Mostly, i worked on adding the container to the icons. It was originally on the footer, but it was limiting the width so I have the icons centered but the width of the darker grey footer spans most widths. On smaller screens, using media queries, there will be equivalent margin as all the other sections.

I will compare this to other pages to see if I should change it.

Maybe look at reuters, or Medium?

Also removed blue-sticky-header because it had poor naming