JavaScript List Project !!!
700 lines of code, soon to balloon into 7000+ lines of code with entire definition list.

I am applying individual functions to each terminology because I don't know how to use variables for this.

The right key isn't specific but it is absolute positioned, so this is not good responsively.

Scroll to and single-click button works at least!

Need to figure out how to add an event listener for onclick that makes any list item disappear.