Updated organized list
Removed an article element from the about area, not sure how it effects this semantically for a11y purposes.

However, my about page now has a centered organized list.

Also, I fixed the cell phone issue. It was at 300% zoom, but not sure if it not looking good on mobile at that % is an issue or not, again, for a11y.

Next focus should be making the form work on jonnovative.biz.